--- a/pkg/logql/log/labels.go
+++ b/pkg/logql/log/labels.go
@@ -240,7 +240,7 @@ func (b *LabelsBuilder) Set(n, v string) *LabelsBuilder {
 
 	// Sometimes labels are set and later modified. Only record
 	// each label once
-	b.parserKeyHints.RecordExtracted()
+	b.parserKeyHints.RecordExtracted(n)
 	return b
 }