@MasslessParticle 2023-03-10
@jeschkies I chatted with @slim-bean and @owen-d  and decided that this is worth the breaking change for something that ought to be an edge case.